+++ Source/Krypton Components/Krypton.Toolkit/Palette Builtin/Office 2010/Extra Themes/Schemes/PaletteOffice2010BlueLightMode_BaseScheme.cs
@@ -138,6 +138,8 @@ public sealed class PaletteOffice2010BlueLightMode_BaseScheme : KryptonColorSche
     public override Color RibbonGroupFrameInside3          { get; set; } = Color.FromArgb(214, 228, 246);
     public override Color RibbonGroupFrameInside4          { get; set; } = Color.FromArgb(227, 236, 248);
     public override Color RibbonGroupCollapsedText         { get; set; } = Color.FromArgb(21, 66, 139);
+    public override Color RibbonGroupTextTracking          { get; set; } = GlobalStaticValues.EMPTY_COLOR;
+    public override Color ButtonTextTracking               { get; set; } = GlobalStaticValues.EMPTY_COLOR;
     public override Color RibbonGroupButtonText            { get; set; } = GlobalStaticValues.EMPTY_COLOR; // missing value
     public override Color AlternatePressedBack1            { get; set; } = Color.FromArgb(118, 153, 200);
     public override Color AlternatePressedBack2            { get; set; } = Color.FromArgb(184, 215, 253);